requirements.txt
pandas==1.4.1
plotly==5.6.0
scipy==1.6.1
sqlalchemy>=1.4.31
pyodbc>=4.0.32
pydantic==1.9.0
pyyaml==6.0